Fix bizarro displayName bug
